Scoreboard + score bug fixed
Scoreboard displayed no longer displays to people with the same name. The one with the most experience points remain(or kills if experience points are the same).

You can not have a name that is more than 10 characters long.

Game data can now be saved, I forgot to implement them for some reason