Highscore/Scoreboard fixed
In MainMenu the newest score was run over by saving it again when quitting in the MainMenu. Fixed it by moving saving to the PlayingScene. But I still needed the LoadManager in the MainMenu. So instead of just adding the GameData Json reading stuff to the Scoreboard class I instead told it to not save its data if it was currently in the MainMenu.